+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Support portable `Update(Note)` activities.
 - Update cached ActivityPub object when processing `Update(Object)` activity.
 - Reject `Create` activity if actor doesn't match object owner.
+- Allow followers-only self-replies to followers-only posts.
 
 ### Fixed